% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertxls.R
\name{convert.xls}
\alias{convert.xls}
\title{Convert XLS to CSV}
\usage{
convert.xls(filename, out.dir = getwd())
}
\arguments{
\item{filename}{the Excel file to be split into .csv files.}

\item{out.dir}{the directory that the .csv files are to be written to. Default is current working directory.}
}
\description{
This function converts a .xlsx or .xls file into .csv files to the chosen directory and merges the sheets together into a single data frame. 
This function is dependent on installing the gnumeric package via command line or through some other method. See the resource for how to install on Linux.
Credits: Chris Wallace
Adapted from: https://cwcode.wordpress.com/2014/04/11/reading-xlsx-files-into-r-quickly/
Resource: https://linuxconfig.org/converting-xlsx-excel-format-files-to-csv-on-linux
}
\examples{
convert.xls(filename="Spreadsheet.xlsx") # writes out spreadsheet files as CSV to working directory
}
\keyword{convert}
\keyword{csv}
\keyword{excel}
\keyword{spreadsheet}
\keyword{xls}
\keyword{xlsx}
