% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfcat.R
\name{dfcat}
\alias{dfcat}
\title{Concatenate a directory of CSV files into one Data Frame}
\usage{
dfcat(directory = getwd())
}
\arguments{
\item{dir}{the directory containing CSV files that are to be concatenated. Default is current working directory.}
}
\description{
This function allows you to read in a directory of CSV files with the same (or different) headers and merge them into one big data frame.
Credits: Hadley Wickham, flodel
https://stackoverflow.com/questions/23190280/issue-in-loading-multiple-csv-files-into-single-dataframe-in-r-using-rbind
https://stackoverflow.com/questions/9807945/consolidating-data-frames-in-r
}
\examples{
dfcat("~/Desktop/CSV Files") # reads in CSV files from the specified directory and merges them into one big data frame.
}
\keyword{concatenate}
\keyword{csv}
\keyword{data}
\keyword{df}
\keyword{frame}
\keyword{list}
