% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left.R
\name{left}
\alias{left}
\title{Left function}
\usage{
left(string, char)
}
\arguments{
\item{string}{a character string or valid expression.}

\item{char}{number of characters to keep starting from the first character of the string.}
}
\description{
This function allows you to trim a character string, starting from the beginning. Equivalent to the 
LEFT() function in Excel.
Credit: Andrea Cirillo 
https://stackoverflow.com/questions/15897236/whats-the-equivalent-to-excels-left-and-right-in-r
}
\examples{
left("moose",3) # returns "moo"
}
\keyword{left}
