% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{Multi Global Substitution}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{Input vector with strings that are to be replaced.}

\item{replacement}{Input vector with the replacement strings.}

\item{x}{a character vector for which replacement is sought.}
}
\description{
This function allows you to make multiple global substitutions.
Credit: Theodore Lytras
https://stackoverflow.com/questions/15253954/replace-multiple-letters-with-accents-with-gsub
}
\examples{
mgsub(c("dog", "tree"), c("cat", "plant"), "The dog and the tree") # returns "The cat and the plant"
}
\keyword{gsub}
\keyword{multi}
\keyword{multiple}
\keyword{substitution}
