% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proper.R
\name{proper}
\alias{proper}
\title{Proper function}
\usage{
proper(x)
}
\arguments{
\item{string}{a character string or valid expression.}
}
\description{
This function allows you to capitalize the first letter of a word and lowercase subsequent letters. Equivalent to the PROPER() function in Excel. 
Credit: Matthew Plourde
https://stackoverflow.com/questions/24956546/capitalizing-letters-r-equivalent-of-excel-proper-function
}
\examples{
proper("moose") # returns "Moose"
}
\keyword{proper}
