% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right.R
\name{right}
\alias{right}
\title{Right function}
\usage{
right(string, char)
}
\arguments{
\item{string}{a character string or valid expression.}

\item{char}{number of characters to keep starting from the last character of the string.}
}
\description{
This function allows you to trim a character string, starting from the end. Equivalent to the 
RIGHT() function in Excel. 
Credit: Andrea Cirillo 
https://stackoverflow.com/questions/15897236/whats-the-equivalent-to-excels-left-and-right-in-r
}
\examples{
left("dragonfly",3) # returns "fly"
}
\keyword{right}
