% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split2csv.R
\name{split2csv}
\alias{split2csv}
\title{Split List to CSV by Factor}
\usage{
split2csv(x, f, directory = getwd())
}
\arguments{
\item{x}{the vector or data frame containing values to be divided into groups.}

\item{f}{a ‘factor’ in the sense that as.factor(f) defines the grouping, or a list of such factors in which case their interaction is used for the grouping.}

\item{directory}{the directory that the CSV files will be written out to. Default is current working directory. 
Credit: agstudy, Tyler Rinker}
}
\description{
This function allows you to split a data frame into a list of data frames by a chosen factor and write them out as separate CSV files in the chosen directory.
}
\examples{
split2csv(mtcars, 'carb') # writes CSV files to the current working directory using the element name as filenames.
}
\keyword{csv}
\keyword{data}
\keyword{df}
\keyword{factor}
\keyword{frame}
\keyword{list}
\keyword{split}
