% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitcsv.R
\name{splitcsv}
\alias{splitcsv}
\title{Split List to CSV}
\usage{
splitcsv(list)
}
\arguments{
\item{list}{the list containing data frames that are to be split into CSV files.
Credit: agstudy}
}
\description{
This function allows you to split a list of data frames into separate CSV files.
}
\examples{
splitcsv(df_list) # writes CSV files to directory using the element name as filenames.
}
\keyword{csv}
\keyword{data}
\keyword{df}
\keyword{frame}
\keyword{list}
\keyword{split}
